Add documentation to vision service
