+++ Microsoft.Toolkit.Uwp.Services/Microsoft.Toolkit.Uwp.Services.csproj
@@ -60,10 +60,10 @@
     <Compile Include="Services\Bing\BingParser.cs" />
     <Compile Include="Services\Bing\BingService.cs" />
     <Compile Include="Services\CognitiveServices\VisionService.cs" />
-    <Compile Include="Services\CognitiveServices\VisionService.Data.Ocr.cs" />
-    <Compile Include="Services\CognitiveServices\VisionService.Data.Tag.cs" />
-    <Compile Include="Services\CognitiveServices\VisionService.Data.cs" />
-    <Compile Include="Services\CognitiveServices\VisionService.Exceptions.cs" />
+    <Compile Include="Services\CognitiveServices\Data\OcrData.cs" />
+    <Compile Include="Services\CognitiveServices\Data\ImageTags.cs" />
+    <Compile Include="Services\CognitiveServices\Data\RequestExceptionDetails.cs" />
+    <Compile Include="Services\CognitiveServices\VisionServiceException.cs" />
     <Compile Include="Services\CognitiveServices\VisionService.Helper.cs" />
     <Compile Include="Services\CognitiveServices\VisionServiceJsonHelper.cs" />
     <Compile Include="Services\Facebook\FacebookAlbum.cs" />
