--- Microsoft.Toolkit.Uwp.Services/Services/CognitiveServices/VisionService.Data.Tag.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-
-namespace Microsoft.Toolkit.Uwp.Services.CognitiveServices
-{
-    public class ImageTags
-    {
-        [JsonProperty("tags")]
-        public Tag[] Tags { get; set; }
-
-        [JsonProperty("requestId")]
-        public string RequestId { get; set; }
-
-        [JsonProperty("metadata")]
-        public Metadata Metadata { get; set; }
-
-        /// <inheritdoc/>
-        public override string ToString()
-        {
-            return VisionServiceJsonHelper.Stringify(this);
-        }
-    }
-
-    public class Metadata
-    {
-        [JsonProperty("width")]
-        public int Width { get; set; }
-
-        [JsonProperty("height")]
-        public int Height { get; set; }
-
-        [JsonProperty("format")]
-        public string Format { get; set; }
-    }
-
-    public class Tag
-    {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("confidence")]
-        public float Confidence { get; set; }
-    }
-}
