--- Microsoft.Toolkit.Uwp.Services/Services/CognitiveServices/VisionService.Data.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-
-namespace Microsoft.Toolkit.Uwp.Services.CognitiveServices
-{
-    public class RequestExceptionDetails
-    {
-        [JsonProperty("code")]
-        public string Code { get; set; }
-
-        [JsonProperty("requestId")]
-        public string RequestId { get; set; }
-
-        [JsonProperty("message")]
-        public string Message { get; set; }
-    }
-}
